// src/lib/types/vehicle.ts
export interface VehiclePublic {  
  brand: string;  
  model: string;  
  vin: string;  
  cost: number;  
  plate: string;  
  purchaseDate: string;  
  photoUrl?: string;  
  registrationDate: string;  
}  
  
export interface Vehicle {  
  brand: string;  
  model: string;  
  vin: string;  
  cost: number;  
  plate: string;  
  purchaseDate: string;  
  registrationDate: string;
}  
  